fix(deviceio): load the profile query optionally, and move the mapping to cloudxr
Two review findings.

loadExtensionFunction throws when a symbol is missing, so declaring
xrGetCurrentInteractionProfile2NV optional and null-checking it at the call
site was not enough: ActionContextFunctions::load would throw first and take
controller tracker construction with it on any runtime lacking the function.
Look the symbol up directly and leave the pointer null instead, which is what
the null-check already expected. CloudXR provides the function, so this was
latent rather than observed.

identify_headset and its table move from isaacteleop.deviceio to
isaacteleop.cloudxr: deviceio is a C++ module plus its bindings, and this is
neither. The mapping is also specific to what a CloudXR runtime reports, so
cloudxr is where a reader would look for it.
